better stats.  VOP_SCANHEAPS and COP_EXT_HEAPINFO added to valhallaComm --mg
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@7624 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/group.c
@@ -256,8 +256,8 @@ char *GroupName(long address, int isCode)
 
   while (current){
     if (isCode){
-      if ((GroupCodeStart(current) <= address) &&
-	  (address <= GroupCodeEnd(current))){
+      if ((GroupCodeStart(current) <= (unsigned long)address) &&
+	  ((unsigned long)address <= GroupCodeEnd(current))){
 	TRACE_GROUP(fprintf (output, "GroupName: returns "));
 	TRACE_GROUP(fprintf (output, "%s\n", NameOfGroupMacro (current)));
 	return NameOfGroupMacro (current);
